refactor(alerts): split notification preferences
Move alert notification preference serialization and AlertSettings conversion into a focused module while preserving alert_notification_system imports.

Tested: pytest tests/test_alert_notification_system.py tests/test_alert_manager.py tests/test_alert_lifecycle.py tests/test_notification_test.py tests/test_app_notification_activation.py; ruff check src tests scripts; pyright.

Confidence: high
